Suppressed the auto-scrolling on click and fixed initial-jump issues.
For the initial auto-scrolling to the selected route in location.hash:
 - Fixed typo in the calculation of the offset-top position.
 - Fixed requesting the selected route that has special chars (e.g. "/content/{page}")
 - Fixed for refreshes with F5. It needs a setTimeout workaround to get that working in Webkit.
    It needs also in Firefox a other dom element to fire scrollTop at.

Fixed also some old .delegation calls, which are deprecated in jQuery 1.7.

Why suppressing auto-scrolling on click:
When clicking through all methods it's very annoying when the browser always jumps to the clicked method,
especially when you go through all methods from bottom to up. This jump is unexpected and disturbing.