Add default value from reflection
This change provides a means to add the default value from the
reflection of a property should the default value not have been set
previously via a `default` parameter of an attribute or annotation.

Also we check before setting anything to nullable whether a default has
been set either via the reflection of via the default property of the
schema. In those cases we do not add the property to the required
properties as it does have a default