feat(#2297): (Symfony 7.1) Add MapRequestPayload array parameter handling (#2298)
| Q | A |

|---------------|---------------------------------------------------------------------------------------------------------------------------|
| Bug fix? | no |
| New feature? | yes <!-- please update src/**/CHANGELOG.md files --> |
| Deprecations? | no <!-- please update UPGRADE-*.md and
src/**/CHANGELOG.md files --> |
| Issues | Fix #2297 <!-- prefix each issue number with "Fix #", no need
to create an issue if none exists, explain below instead --> |

With Symfony 7.1, a new `type` parameter has been introduced for
`MapRequestPayload` that allows array Controller parameters to be
resolved. This currently leads to an error when the Describer tries to
register a new Model, because the parameter type `array` is used in this
case. To resolve the issue, the given `type` parameter form the
`MapRequestPayload` should be used instead.

I tried to test this as gracefully as possible. Any more elegant
suggestions are welcome!

---------

Co-authored-by: Maximilian Zumbansen <maximilian.zumbansen@qossmic.com>
Co-authored-by: djordy <djordy.koert@yoursurprise.com>