feat(2056): support opt out of JMS serializer usage (#2342)
| Q | A |

|---------------|---------------------------------------------------------------------------------------------------------------------------|
| Bug fix? | no |
| New feature? | yes <!-- please update src/**/CHANGELOG.md files --> |
| Deprecations? | no <!-- please update UPGRADE-*.md and
src/**/CHANGELOG.md files --> |
| Issues | Fix #2056 #2341 <!-- prefix each issue number with "Fix #",
no need to create an issue if none exists, explain below instead --> |

<!--
Replace this notice by a description of your feature/bugfix.
This will help reviewers and should be a good start for the
documentation.

Additionally:
 - Always add tests and ensure they pass.
- For new features, provide some code snippets to help understand usage.
-->
Always registers `JMSModelDescriber` when the bundle is active,
independently of the configured global `models.use_jms` option.
If `models.use_jms` is set, then `JMSModelDescriber` will be added with
a priority of 50, just as before. If `models.use_jms` is not set, it
will be added with -50 so that `ObjectModelDescriber` is used instead.
This should not lead to any behavior change.

Adds support to `options.useJms` in the model by early exiting in either
`JMSModelDescriber` or `ObjectModelDescriber` when it is either
explicitly `false` or `true`. Together with the priorities of the
describer this allows to have either JMS or Symfony serializer as
default and switch per model if necessary.

In https://github.com/nelmio/NelmioApiDocBundle/issues/2341 I played
around with this in an app and it worked as expected.

(sorry btw, I'm a fan of rebase + force-push and realized too late that
mentioning the issue in the first commit is a bad idea then because
GitHub spams whenever I push..)

Closes https://github.com/nelmio/NelmioApiDocBundle/issues/2056
Closes https://github.com/nelmio/NelmioApiDocBundle/issues/2341

---------

Co-authored-by: djordy <djordy.koert@yoursurprise.com>