fix: pass through serializationContext into models where possible (#2344)
| Q | A |

|---------------|---------------------------------------------------------------------------------------------------------------------------|
| Bug fix? | yes |
| New feature? | o <!-- please update src/**/CHANGELOG.md files --> |
| Deprecations? | no <!-- please update UPGRADE-*.md and
src/**/CHANGELOG.md files --> |
| Issues | Fix #... <!-- prefix each issue number with "Fix #", no need
to create an issue if none exists, explain below instead --> |

<!--
Replace this notice by a description of your feature/bugfix.
This will help reviewers and should be a good start for the
documentation.

Additionally:
 - Always add tests and ensure they pass.
- For new features, provide some code snippets to help understand usage.
-->
I noticed that information I'm adding into a JMS serializer model gets
lost inside it because `serializationContext` is not passed through
everywhere, but it most likely should? the thing I slightly dislike
though is that it _could_ include `groups` and in the scope of
`JMSModelDescriber` where groups are determined in another way and
overwritten I have to unset that key from the serializationContext to
avoid side-effects. WDYT @DjordyKoert?