feat: support multiple authorisations to be stored and restored (#2311)
| Q | A |

|---------------|---------------------------------------------------------------------------------------------------------------------------|
| Bug fix? | maybe |
| New feature? | maybe <!-- please update src/**/CHANGELOG.md files -->
|
| Deprecations? | no <!-- please update UPGRADE-*.md and
src/**/CHANGELOG.md files --> |
| Issues | - <!-- prefix each issue number with "Fix #", no need to
create an issue if none exists, explain below instead --> |

<!--
Replace this notice by a description of your feature/bugfix.
This will help reviewers and should be a good start for the
documentation.

Additionally:
 - Always add tests and ensure they pass.
- For new features, provide some code snippets to help understand usage.
-->

I have an API description which has multiple authorisation options,
which is supported fine by the Swagger interfaces. However, the
store/restore functionality always overrides the previous stored value,
so only the last input is restored on page load. This change solves
that.

As far as I know everything I used should work fine on relevant
browsers.
  - Object.keys: https://caniuse.com/mdn-javascript_builtins_object_keys
  - forEach: https://caniuse.com/mdn-javascript_builtins_array_foreach

I do not believe that the build failures are related to this PR....

---------

Co-authored-by: Djordy Koert <djordy.koert@gmail.com>