feat: add support for name based serialisation of JMS enums (#2355)
| Q | A |

|---------------|---------------------------------------------------------------------------------------------------------------------------|
| Bug fix? | no |
| New feature? | yes <!-- please update src/**/CHANGELOG.md files --> |
| Deprecations? | no <!-- please update UPGRADE-*.md and
src/**/CHANGELOG.md files --> |
| Issues | - <!-- prefix each issue number with "Fix #", no need to
create an issue if none exists, explain below instead --> |

<!--
Replace this notice by a description of your feature/bugfix.
This will help reviewers and should be a good start for the
documentation.

Additionally:
 - Always add tests and ensure they pass.
- For new features, provide some code snippets to help understand usage.
-->

This PR add a method to detect when the developer has annotated their
property with a JMS type that specifies that the backed enum needs to be
serialized with its name instead of value. I needed to use the
serialization context as the model options are not included in the model
hash, while I do need to generate a new unique description. I've chosen
to append `Name` to the resulting Model name.

Depends on #2372

(cherry picked from commit 9ee5f586213fc77c9adc56bd59cc9229a9da2527)