chore: allow zircote/swagger-php v5 (#2420)
## Description

zircote/swagger-php v5 has been released, and seems to only require
minimal updates. See
https://github.com/zircote/swagger-php/blob/master/docs/guide/migrating-to-v5.md.

I did need to migrate the current processor implementation to the one
that is using the new processor pipeline introduced with 4.10.0, which
is why I bumped the minimum constraint. In order to not mess up
dependency injection, I was required to wrap the Generator class in
order to allow adding additional processors using the existing compiler
pass.

## What type of PR is this? (check all applicable)
- [ ] Bug Fix
- [x] Feature
- [ ] Refactor
- [ ] Deprecation
- [ ] Breaking Change
- [ ] Documentation Update
- [ ] CI

## Checklist
- [ ] I have made corresponding changes to the documentation (`docs/`)
- [ ] I have made corresponding changes to the changelog
(`CHANGELOG.md`)

---------

Co-authored-by: djordy <djordy.koert@yoursurprise.com>